Refactor config. framework feature definition
In spite of using the sign_in_token option as an example option,
this feature definition does not test if the Devise `sign_in` method
gets the correct arguments, but only tests that the configuration
framework allows to store settings as expected.

As they were written, the scenarii were a little missleading.
In fact, I was wrong when stating that c983d17 did fix the issue #19.