Add RSpec to the gem (not only the dummy app)
The Cucumber features ensure the gem behaves as expected, however
the increasing complexity of the gem internals makes me wish to
be able to write some unit tests.

The idea is to describe the lib/ components behaviour with RSpec,
each on their own; and to ensure they work well together with Cucumber.

That should also make contributions easier: the features testing
is becoming a serious bottleneck and the complexity of the Cucumber
setup doesn't help at all. Some unit tests should help reducing that
complexity in the long term. Here we go! :)