Require ActiveRecord as development dependency (part 1)
Anyone who would not want to install ActiveRecord could modify
the configuration to remove it from the adapters list.

Example:

    # config/initializers/simple_token_authentication.rb

    SimpleTokenAuthentication.configure do |config|
      config.adapters = [] # was ['active_record']
    end

Note: I temporary downgraded the ActiveRecord version to trick Travis-CI.

The build fails when the changes of the `Gemfile/*.gemspec` don't
match those of the `Gemfile.lock`, and just moving the ActiveRecord
dependency without modifying it seems to highly confuse the builder.