Fix Neo4j adapter loading issues
The configuration can't be modified from the Rails app for now.
That's to say that support for new adapters could easily be added
to forks, but not yet on a per-app basis through initializers and
custom adpters.

Expectation: models created from the Neo4j base class for models
shoud now be able to act as token authenticatables and be used
in a token authentication context.