Minor add adapter interface spec
The shared specs for the 'adapter' interface provide a better
help if they spot the methods which miss an override.

Adding this spec makes the shared specs no longer usable with
the `Adapter` module itself. Some more caution will be needed
to ensure both the module and the shared specs stay in sync, but
the module should change less often than new adapters will be added.