Temporary remove token auth. hooks when using Grape adapter
This commit is temporary. It should not be merged as is because
the condition it relies on to disable the hooks setting is aproximative.

Example of corner case: if for some reason a Grape user defined the
`before_filter` method or is using `ActiveController#before_filter` and
is calling `authenticate_user_from_token_or_fallback` (as supposed to),
then the hook will be set and the method will result being called twice.