Fix remove unnecessary Ruby 2.0 requirment
The 'tins' gem is an Inch dependency. Since v1.7.0, it requires
explicitely Ruby 2.

Simple Token Authentication used to support ruby 1.9.3. Even if I don't
think anybody is still using it, I prefer keeping that question for
a major release.

Constraining the version of an inherited dependency is not obvious,
I think the Gemfile is not a bad place to do it.

See
https://github.com/flori/tins/commit/9ebb625d8a987bd54526006441ff97bc14c30e53