Temporary set regular dependencies version constraints
Remove temporary allowed failures for Ruby 1.9.3 and 2.2.0
Minor add version constraint on development dependency

The previous commits ensured that the Rails 5 support was fine.
Since pre-release versions cannot be installed with the regular
version constraints, the regression testing must be performed
independently.

It will be possible to test both cases without changes as soon as
Rails 5 is released.