Remove indirect dependencies from gemspec
As a developer
Given a Ruby 2 app which depends on Simple Token Authentication
In order to use the latest version of tins
I want Simple Token Authentication not to constrain the 'tins' version
unnecessarily

Given a Ruby 1.9.3 app which depends on Simple Token Authentication
In order to make sure the build doesn't break unnecessarily
I want the 'tins' version to be constrained in test environment

That is also true for mime-types.

Note: I thought that could be achieved using Bundler groups instead
of Appraisal, but it cannot.
See http://yehudakatz.com/2010/05/09/the-how-and-why-of-bundler-groups