Prevent Mongoid adapter from loading when unsupported
Unless I'm wrong, no version of Mongoid is currently compatible with Rails 6.

As far as I know at the time of writing, the XML serializer requirement
is currently unmet, and I think it is a reasonable flag to use in order
to load the adpater only when usable.

Note: the syntax of the condition must be compatible with Ruby 1.9.3